LayoutTests/storage/indexeddb/factory-deletedatabase-interactions-expected.txt

Test the deleteDatabase call and its interaction with open/setVersion

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;
PASS Boolean(indexedDB && IDBCursor && IDBDatabase && IDBDatabaseError && IDBDatabaseException && IDBFactory && IDBIndex && IDBKeyRange && IDBObjectStore && IDBRequest && IDBTransaction) is true


TEST: deleteDatabase blocked on open handles
self.dbname = 'test1'; self.ver = 1; self.steps = []
'h.open'
'h.open.onsuccess'
'deleteDatabase()'
'h.onversionchange'
    h closing, but not immediately
'deleteDatabase().onblocked'
'h.close'
'deleteDatabase().onsuccess'
PASS self.steps.toString() is "h.open,h.open.onsuccess,deleteDatabase(),h.onversionchange,deleteDatabase().onblocked,h.close,deleteDatabase().onsuccess"

TEST: deleteDatabase not blocked when handles close immediately
self.dbname = 'test2'; self.ver = 1; self.steps = []
'h.open'
'h.open.onsuccess'
'deleteDatabase()'
'h.onversionchange'
    h closing immediately
'h.close'
'deleteDatabase().onblocked'
'deleteDatabase().onsuccess'
NOTE: Will FAIL with extra bogus deleteDatabase().onblocked step; https://bugs.webkit.org/show_bug.cgi?id=71130
FAIL self.steps.toString() should be h.open,h.open.onsuccess,deleteDatabase(),h.onversionchange,h.close,deleteDatabase().onsuccess. Was h.open,h.open.onsuccess,deleteDatabase(),h.onversionchange,h.close,deleteDatabase().onblocked,deleteDatabase().onsuccess.

TEST: deleteDatabase is delayed if a VERSION_CHANGE transaction is running
self.dbname = 'test3'; self.ver = 1; self.steps = []
'h.open'
'h.open.onsuccess'
'h.setVersion'
'deleteDatabase()'
'h.setVersion.onsuccess'
'h.setVersion.transaction-complete'
'h.onversionchange'
    h closing, but not immediately
'deleteDatabase().onblocked'
'h.close'
'deleteDatabase().onsuccess'
PASS self.steps.toString() is "h.open,h.open.onsuccess,h.setVersion,deleteDatabase(),h.setVersion.onsuccess,h.setVersion.transaction-complete,h.onversionchange,deleteDatabase().onblocked,h.close,deleteDatabase().onsuccess"
PASS successfullyParsed is true

TEST COMPLETE
